+++ layers/GatoGraphQLForWP/plugins/gatographql/src/ModuleResolvers/Extensions/BundleExtensionModuleResolver.php
@@ -9,10 +9,17 @@ use GatoGraphQL\GatoGraphQL\PluginApp;
 
 class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolver
 {
-    public const ALL_EXTENSIONS = Plugin::NAMESPACE . '\\bundle-extensions\\all-extensions';
-    public const APPLICATION_GLUE_AND_AUTOMATOR = Plugin::NAMESPACE . '\\bundle-extensions\\application-glue-and-automator';
-    public const CONTENT_TRANSLATION = Plugin::NAMESPACE . '\\bundle-extensions\\content-translation';
-    public const PUBLIC_API = Plugin::NAMESPACE . '\\bundle-extensions\\public-api';
+    public const ALL_IN_ONE_TOOLBOX_FOR_WORDPRESS = Plugin::NAMESPACE . '\\bundle-extensions\\all-in-one-toolbox-for-wordpress';
+    public const AUTOMATED_CONTENT_TRANSLATION_AND_SYNC_FOR_WORDPRESS_MULTISITE = Plugin::NAMESPACE . '\\bundle-extensions\\automated-content-translation-and-sync-for-wordpress-multisite';
+    public const BETTER_WORDPRESS_WEBHOOKS = Plugin::NAMESPACE . '\\bundle-extensions\\better-wordpress-webhooks';
+    public const EASY_WORDPRESS_BULK_TRANSFORM_AND_UPDATE = Plugin::NAMESPACE . '\\bundle-extensions\\easy-wordpress-bulk-transform-and-update';
+    public const PRIVATE_GRAPHQL_SERVER_FOR_WORDPRESS = Plugin::NAMESPACE . '\\bundle-extensions\\private-graphql-server-for-wordpress';
+    public const RESPONSIBLE_WORDPRESS_PUBLIC_API = Plugin::NAMESPACE . '\\bundle-extensions\\responsible-wordpress-public-api';
+    public const SELECTIVE_CONTENT_IMPORT_EXPORT_AND_SYNC_FOR_WORDPRESS = Plugin::NAMESPACE . '\\bundle-extensions\\selective-content-import-export-and-sync-for-wordpress';
+    public const SIMPLEST_WORDPRESS_CONTENT_TRANSLATION = Plugin::NAMESPACE . '\\bundle-extensions\\simplest-wordpress-content-translation';
+    public const TAILORED_WORDPRESS_AUTOMATOR = Plugin::NAMESPACE . '\\bundle-extensions\\tailored-wordpress-automator';
+    public const UNHINDERED_WORDPRESS_EMAIL_NOTIFICATIONS = Plugin::NAMESPACE . '\\bundle-extensions\\unhindered-wordpress-email-notifications';
+    public const VERSATILE_WORDPRESS_REQUEST_API = Plugin::NAMESPACE . '\\bundle-extensions\\versatile-wordpress-request-api';
 
     /**
      * @return string[]
@@ -20,10 +27,17 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
     public function getModulesToResolve(): array
     {
         return [
-            self::ALL_EXTENSIONS,
-            self::APPLICATION_GLUE_AND_AUTOMATOR,
-            self::CONTENT_TRANSLATION,
-            self::PUBLIC_API,
+            self::ALL_IN_ONE_TOOLBOX_FOR_WORDPRESS,
+            self::AUTOMATED_CONTENT_TRANSLATION_AND_SYNC_FOR_WORDPRESS_MULTISITE,
+            self::BETTER_WORDPRESS_WEBHOOKS,
+            self::EASY_WORDPRESS_BULK_TRANSFORM_AND_UPDATE,
+            self::PRIVATE_GRAPHQL_SERVER_FOR_WORDPRESS,
+            self::RESPONSIBLE_WORDPRESS_PUBLIC_API,
+            self::SELECTIVE_CONTENT_IMPORT_EXPORT_AND_SYNC_FOR_WORDPRESS,
+            self::SIMPLEST_WORDPRESS_CONTENT_TRANSLATION,
+            self::TAILORED_WORDPRESS_AUTOMATOR,
+            self::UNHINDERED_WORDPRESS_EMAIL_NOTIFICATIONS,
+            self::VERSATILE_WORDPRESS_REQUEST_API,
         ];
     }
 
@@ -31,10 +45,17 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
     {
         $placeholder = \__('“%s” Bundle', 'gatographql');
         return match ($module) {
-            self::ALL_EXTENSIONS => sprintf($placeholder, \__('All Extensions', 'gatographql')),
-            self::APPLICATION_GLUE_AND_AUTOMATOR => sprintf($placeholder, \__('Application Glue & Automator', 'gatographql')),
-            self::CONTENT_TRANSLATION => sprintf($placeholder, \__('Content Translation', 'gatographql')),
-            self::PUBLIC_API => sprintf($placeholder, \__('Public API', 'gatographql')),
+            self::ALL_IN_ONE_TOOLBOX_FOR_WORDPRESS => sprintf($placeholder, \__('All in One Toolbox for WordPress', 'gatographql')),
+            self::AUTOMATED_CONTENT_TRANSLATION_AND_SYNC_FOR_WORDPRESS_MULTISITE => sprintf($placeholder, \__('Automated Content Translation & Sync for WordPress Multisite', 'gatographql')),
+            self::BETTER_WORDPRESS_WEBHOOKS => sprintf($placeholder, \__('Better WordPress Webhooks', 'gatographql')),
+            self::EASY_WORDPRESS_BULK_TRANSFORM_AND_UPDATE => sprintf($placeholder, \__('Easy WordPress Bulk Transform & Update', 'gatographql')),
+            self::PRIVATE_GRAPHQL_SERVER_FOR_WORDPRESS => sprintf($placeholder, \__('Private GraphQL Server for WordPress', 'gatographql')),
+            self::RESPONSIBLE_WORDPRESS_PUBLIC_API => sprintf($placeholder, \__('Responsible WordPress Public API', 'gatographql')),
+            self::SELECTIVE_CONTENT_IMPORT_EXPORT_AND_SYNC_FOR_WORDPRESS => sprintf($placeholder, \__('Selective Content Import, Export & Sync for WordPress', 'gatographql')),
+            self::SIMPLEST_WORDPRESS_CONTENT_TRANSLATION => sprintf($placeholder, \__('Simplest WordPress Content Translation', 'gatographql')),
+            self::TAILORED_WORDPRESS_AUTOMATOR => sprintf($placeholder, \__('Tailored WordPress Automator', 'gatographql')),
+            self::UNHINDERED_WORDPRESS_EMAIL_NOTIFICATIONS => sprintf($placeholder, \__('Unhindered WordPress Email Notifications', 'gatographql')),
+            self::VERSATILE_WORDPRESS_REQUEST_API => sprintf($placeholder, \__('Versatile WordPress Request API', 'gatographql')),
             default => $module,
         };
     }
@@ -42,10 +63,17 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
     public function getDescription(string $module): string
     {
         return match ($module) {
-            self::ALL_EXTENSIONS => \__('All of Gato GraphQL extensions, in a single plugin. As new extensions are created and released, they will also be added to this bundle.', 'gatographql'),
-            self::APPLICATION_GLUE_AND_AUTOMATOR => \__('Keep content in sync, help migrate websites, send notifications, interact with 3rd-party services and APIs, create automation workflows, and more.', 'gatographql'),
-            self::CONTENT_TRANSLATION => \__('Translate content via the Google Translate API, even within the deep structure of (Gutenberg) blocks.', 'gatographql'),
-            self::PUBLIC_API => \__('Expose your public APIs in a secure manner, make them faster through caching, leverage tools to access data, and evolve the schema.', 'gatographql'),
+            self::ALL_IN_ONE_TOOLBOX_FOR_WORDPRESS => \__('Achieve all superpowers: All of Gato GraphQL extensions, in a single plugin', 'gatographql'),
+            self::AUTOMATED_CONTENT_TRANSLATION_AND_SYNC_FOR_WORDPRESS_MULTISITE => \__('Automatically create a translation of a newly-published post using the Google Translate API, for every language site on a WordPress multisite', 'gatographql'),
+            self::BETTER_WORDPRESS_WEBHOOKS => \__('Easily create webhooks to process incoming data from any source or service using advanced tools, directly within the wp-admin', 'gatographql'),
+            self::EASY_WORDPRESS_BULK_TRANSFORM_AND_UPDATE => \__('Transform hundreds of posts with a single operation (replacing strings, adding blocks, adding a thumbnail, and more), and store them again on the DB', 'gatographql'),
+            self::PRIVATE_GRAPHQL_SERVER_FOR_WORDPRESS => \__('Use GraphQL to power your application (blocks, themes and plugins), internally fetching data without exposing a public endpoint', 'gatographql'),
+            self::RESPONSIBLE_WORDPRESS_PUBLIC_API => \__('Enhance your public APIs with additional layers of security, speed, power, schema evolution and control', 'gatographql'),
+            self::SELECTIVE_CONTENT_IMPORT_EXPORT_AND_SYNC_FOR_WORDPRESS => \__('Import hundreds of records into your WordPress site from another site or service (such as Google Sheets), and selectively export entries to another site', 'gatographql'),
+            self::SIMPLEST_WORDPRESS_CONTENT_TRANSLATION => \__('Translate your content into over 130 languages using the Google Translate API, without adding extra tables or inner joins to the DB', 'gatographql'),
+            self::TAILORED_WORDPRESS_AUTOMATOR => \__('Create workflows to automate tasks (to transform data, automatically caption images, send notifications, and more)', 'gatographql'),
+            self::UNHINDERED_WORDPRESS_EMAIL_NOTIFICATIONS => \__('Send personalized emails to all your users, and notifications to the admin, without constraints on what data can be added to the email', 'gatographql'),
+            self::VERSATILE_WORDPRESS_REQUEST_API => \__('Interact with any external API and cloud service, posting and fetching data to/from them', 'gatographql'),
             default => parent::getDescription($module),
         };
     }
@@ -58,7 +86,7 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
     public function getLogoURL(string $module): string
     {
         return match ($module) {
-            self::ALL_EXTENSIONS => PluginApp::getMainPlugin()->getPluginURL() . 'assets/img/logos/GatoGraphQL-logo-face.png',
+            self::ALL_IN_ONE_TOOLBOX_FOR_WORDPRESS => PluginApp::getMainPlugin()->getPluginURL() . 'assets/img/logos/GatoGraphQL-logo-face.png',
             default => parent::getLogoURL($module),
         };
     }
@@ -69,7 +97,7 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
     public function getBundledExtensionModules(string $module): array
     {
         return match ($module) {
-            self::ALL_EXTENSIONS => [
+            self::ALL_IN_ONE_TOOLBOX_FOR_WORDPRESS => [
                 ExtensionModuleResolver::ACCESS_CONTROL,
                 ExtensionModuleResolver::ACCESS_CONTROL_VISITOR_IP,
                 ExtensionModuleResolver::AUTOMATION,
@@ -97,7 +125,109 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
                 ExtensionModuleResolver::RESPONSE_ERROR_TRIGGER,
                 ExtensionModuleResolver::SCHEMA_EDITING_ACCESS,
             ],
-            self::APPLICATION_GLUE_AND_AUTOMATOR => [
+            self::AUTOMATED_CONTENT_TRANSLATION_AND_SYNC_FOR_WORDPRESS_MULTISITE => [
+                ExtensionModuleResolver::AUTOMATION,
+                ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
+                ExtensionModuleResolver::FIELD_ON_FIELD,
+                ExtensionModuleResolver::FIELD_RESPONSE_REMOVAL,
+                ExtensionModuleResolver::FIELD_TO_INPUT,
+                ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
+                ExtensionModuleResolver::GOOGLE_TRANSLATE,
+                ExtensionModuleResolver::HTTP_CLIENT,
+                ExtensionModuleResolver::INTERNAL_GRAPHQL_SERVER,
+                ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
+                ExtensionModuleResolver::PHP_CONSTANTS_AND_ENVIRONMENT_VARIABLES_VIA_SCHEMA,
+                ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
+                ExtensionModuleResolver::RESPONSE_ERROR_TRIGGER,
+            ],
+            self::BETTER_WORDPRESS_WEBHOOKS => [
+                ExtensionModuleResolver::ACCESS_CONTROL,
+                ExtensionModuleResolver::ACCESS_CONTROL_VISITOR_IP,
+                ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
+                ExtensionModuleResolver::EMAIL_SENDER,
+                ExtensionModuleResolver::FIELD_DEFAULT_VALUE,
+                ExtensionModuleResolver::FIELD_ON_FIELD,
+                ExtensionModuleResolver::FIELD_RESPONSE_REMOVAL,
+                ExtensionModuleResolver::FIELD_TO_INPUT,
+                ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
+                ExtensionModuleResolver::HELPER_FUNCTION_COLLECTION,
+                ExtensionModuleResolver::HTTP_CLIENT,
+                ExtensionModuleResolver::HTTP_REQUEST_VIA_SCHEMA,
+                ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
+                ExtensionModuleResolver::PHP_CONSTANTS_AND_ENVIRONMENT_VARIABLES_VIA_SCHEMA,
+                ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
+                ExtensionModuleResolver::RESPONSE_ERROR_TRIGGER,
+            ],
+            self::EASY_WORDPRESS_BULK_TRANSFORM_AND_UPDATE => [
+                ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
+                ExtensionModuleResolver::FIELD_DEFAULT_VALUE,
+                ExtensionModuleResolver::FIELD_ON_FIELD,
+                ExtensionModuleResolver::FIELD_TO_INPUT,
+                ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
+                ExtensionModuleResolver::HELPER_FUNCTION_COLLECTION,
+                ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
+                ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
+            ],
+            self::PRIVATE_GRAPHQL_SERVER_FOR_WORDPRESS => [
+                ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
+                ExtensionModuleResolver::EMAIL_SENDER,
+                ExtensionModuleResolver::FIELD_DEFAULT_VALUE,
+                ExtensionModuleResolver::FIELD_ON_FIELD,
+                ExtensionModuleResolver::FIELD_RESOLUTION_CACHING,
+                ExtensionModuleResolver::FIELD_RESPONSE_REMOVAL,
+                ExtensionModuleResolver::FIELD_TO_INPUT,
+                ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
+                ExtensionModuleResolver::HELPER_FUNCTION_COLLECTION,
+                ExtensionModuleResolver::HTTP_CLIENT,
+                ExtensionModuleResolver::HTTP_REQUEST_VIA_SCHEMA,
+                ExtensionModuleResolver::INTERNAL_GRAPHQL_SERVER,
+                ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
+                ExtensionModuleResolver::PHP_CONSTANTS_AND_ENVIRONMENT_VARIABLES_VIA_SCHEMA,
+                ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
+                ExtensionModuleResolver::RESPONSE_ERROR_TRIGGER,
+            ],
+            self::RESPONSIBLE_WORDPRESS_PUBLIC_API => [
+                ExtensionModuleResolver::ACCESS_CONTROL,
+                ExtensionModuleResolver::ACCESS_CONTROL_VISITOR_IP,
+                ExtensionModuleResolver::CACHE_CONTROL,
+                ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
+                ExtensionModuleResolver::DEPRECATION_NOTIFIER,
+                ExtensionModuleResolver::FIELD_DEFAULT_VALUE,
+                ExtensionModuleResolver::FIELD_DEPRECATION,
+                ExtensionModuleResolver::FIELD_TO_INPUT,
+                ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
+                ExtensionModuleResolver::LOW_LEVEL_PERSISTED_QUERY_EDITING,
+                ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
+                ExtensionModuleResolver::RESPONSE_ERROR_TRIGGER,
+                ExtensionModuleResolver::SCHEMA_EDITING_ACCESS,
+            ],
+            self::SELECTIVE_CONTENT_IMPORT_EXPORT_AND_SYNC_FOR_WORDPRESS => [
+                ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
+                ExtensionModuleResolver::FIELD_DEFAULT_VALUE,
+                ExtensionModuleResolver::FIELD_ON_FIELD,
+                ExtensionModuleResolver::FIELD_RESOLUTION_CACHING,
+                ExtensionModuleResolver::FIELD_RESPONSE_REMOVAL,
+                ExtensionModuleResolver::FIELD_TO_INPUT,
+                ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
+                ExtensionModuleResolver::HELPER_FUNCTION_COLLECTION,
+                ExtensionModuleResolver::HTTP_CLIENT,
+                ExtensionModuleResolver::HTTP_REQUEST_VIA_SCHEMA,
+                ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
+                ExtensionModuleResolver::PHP_CONSTANTS_AND_ENVIRONMENT_VARIABLES_VIA_SCHEMA,
+                ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
+                ExtensionModuleResolver::RESPONSE_ERROR_TRIGGER,
+            ],
+            self::SIMPLEST_WORDPRESS_CONTENT_TRANSLATION => [
+                ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
+                ExtensionModuleResolver::FIELD_ON_FIELD,
+                ExtensionModuleResolver::FIELD_RESPONSE_REMOVAL,
+                ExtensionModuleResolver::FIELD_TO_INPUT,
+                ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
+                ExtensionModuleResolver::GOOGLE_TRANSLATE,
+                ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
+                ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
+            ],
+            self::TAILORED_WORDPRESS_AUTOMATOR => [
                 ExtensionModuleResolver::AUTOMATION,
                 ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
                 ExtensionModuleResolver::EMAIL_SENDER,
@@ -116,30 +246,31 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
                 ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
                 ExtensionModuleResolver::RESPONSE_ERROR_TRIGGER,
             ],
-            self::CONTENT_TRANSLATION => [
-                ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
+            self::UNHINDERED_WORDPRESS_EMAIL_NOTIFICATIONS => [
+                ExtensionModuleResolver::EMAIL_SENDER,
                 ExtensionModuleResolver::FIELD_ON_FIELD,
+                ExtensionModuleResolver::FIELD_TO_INPUT,
+                ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
+                ExtensionModuleResolver::HELPER_FUNCTION_COLLECTION,
+                ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
+                ExtensionModuleResolver::PHP_CONSTANTS_AND_ENVIRONMENT_VARIABLES_VIA_SCHEMA,
+                ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
+            ],
+            self::VERSATILE_WORDPRESS_REQUEST_API => [
+                ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
+                ExtensionModuleResolver::FIELD_DEFAULT_VALUE,
+                ExtensionModuleResolver::FIELD_ON_FIELD,
+                ExtensionModuleResolver::FIELD_RESOLUTION_CACHING,
                 ExtensionModuleResolver::FIELD_RESPONSE_REMOVAL,
                 ExtensionModuleResolver::FIELD_TO_INPUT,
                 ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
-                ExtensionModuleResolver::GOOGLE_TRANSLATE,
+                ExtensionModuleResolver::HELPER_FUNCTION_COLLECTION,
+                ExtensionModuleResolver::HTTP_CLIENT,
+                ExtensionModuleResolver::HTTP_REQUEST_VIA_SCHEMA,
                 ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
+                ExtensionModuleResolver::PHP_CONSTANTS_AND_ENVIRONMENT_VARIABLES_VIA_SCHEMA,
                 ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
-            ],
-            self::PUBLIC_API => [
-                ExtensionModuleResolver::ACCESS_CONTROL,
-                ExtensionModuleResolver::ACCESS_CONTROL_VISITOR_IP,
-                ExtensionModuleResolver::CACHE_CONTROL,
-                ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
-                ExtensionModuleResolver::DEPRECATION_NOTIFIER,
-                ExtensionModuleResolver::FIELD_DEFAULT_VALUE,
-                ExtensionModuleResolver::FIELD_DEPRECATION,
-                ExtensionModuleResolver::FIELD_TO_INPUT,
-                ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
-                ExtensionModuleResolver::LOW_LEVEL_PERSISTED_QUERY_EDITING,
-                ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
                 ExtensionModuleResolver::RESPONSE_ERROR_TRIGGER,
-                ExtensionModuleResolver::SCHEMA_EDITING_ACCESS,
             ],
             default => [],
         };
@@ -151,11 +282,39 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
     public function getBundledBundleExtensionModules(string $module): array
     {
         return match ($module) {
-            // "All Extensions" bundles all other bundles
-            self::ALL_EXTENSIONS => array_diff(
+            // "All in One Toolbox for WordPress" bundles all other bundles
+            self::ALL_IN_ONE_TOOLBOX_FOR_WORDPRESS => array_diff(
                 $this->getModulesToResolve(),
                 [$module]
             ),
+            self::AUTOMATED_CONTENT_TRANSLATION_AND_SYNC_FOR_WORDPRESS_MULTISITE => [
+                self::SIMPLEST_WORDPRESS_CONTENT_TRANSLATION,
+            ],
+            self::PRIVATE_GRAPHQL_SERVER_FOR_WORDPRESS => [
+                self::SELECTIVE_CONTENT_IMPORT_EXPORT_AND_SYNC_FOR_WORDPRESS,
+                self::EASY_WORDPRESS_BULK_TRANSFORM_AND_UPDATE,
+                self::UNHINDERED_WORDPRESS_EMAIL_NOTIFICATIONS,
+                self::VERSATILE_WORDPRESS_REQUEST_API,
+            ],
+            self::BETTER_WORDPRESS_WEBHOOKS => [
+                self::EASY_WORDPRESS_BULK_TRANSFORM_AND_UPDATE,
+                self::UNHINDERED_WORDPRESS_EMAIL_NOTIFICATIONS,
+                self::VERSATILE_WORDPRESS_REQUEST_API,
+            ],
+            self::TAILORED_WORDPRESS_AUTOMATOR => [
+                self::PRIVATE_GRAPHQL_SERVER_FOR_WORDPRESS,
+                self::SELECTIVE_CONTENT_IMPORT_EXPORT_AND_SYNC_FOR_WORDPRESS,
+                self::EASY_WORDPRESS_BULK_TRANSFORM_AND_UPDATE,
+                self::UNHINDERED_WORDPRESS_EMAIL_NOTIFICATIONS,
+                self::VERSATILE_WORDPRESS_REQUEST_API,
+            ],
+            self::SELECTIVE_CONTENT_IMPORT_EXPORT_AND_SYNC_FOR_WORDPRESS => [
+                self::EASY_WORDPRESS_BULK_TRANSFORM_AND_UPDATE,
+            ],
+            self::VERSATILE_WORDPRESS_REQUEST_API => [
+                self::SELECTIVE_CONTENT_IMPORT_EXPORT_AND_SYNC_FOR_WORDPRESS,
+                self::EASY_WORDPRESS_BULK_TRANSFORM_AND_UPDATE,
+            ],
             default => [],
         };
     }