+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -14,6 +14,18 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 - Added "Request headers" to GraphiQL clients on single public/private endpoint, and custom endpoints
 - Renamed page "Recipes" to "Tutorial", and added settings to hide it
+- Renamed existing bundles:
+  - “Application Glue & Automator” Bundle => “Tailored WordPress Automator” Bundle
+  - “Content Translation” Bundle => “Simplest WordPress Content Translation” Bundle
+  - “Public API” Bundle => “Responsible WordPress Public API” Bundle
+- Added documentation for new bundles:
+  - “Automated Content Translation & Sync for WordPress Multisite” Bundle
+  - “Better WordPress Webhooks” Bundle
+  - “Easy WordPress Bulk Transform & Update” Bundle
+  - “Private GraphQL Server for WordPress” Bundle
+  - “Selective Content Import, Export & Sync for WordPress” Bundle
+  - “Unhindered WordPress Email Notifications” Bundle
+  - “Versatile WordPress Request API” Bundle
 
 ### Fixed